rename ExceptionLogger and ScopedExceptionLogger. Logger is imprecise here because there are expectations of a logger that our objects here does not satisfy (e.g. outputs to a stream, collects loglines, etc.)
