+++ flask_monitoringdashboard/__init__.py
@@ -18,8 +18,8 @@ from flask import Blueprint, g
 
 from flask_monitoringdashboard.core.config import Config, TelemetryConfig
 from flask_monitoringdashboard.core.logger import log
-from flask_monitoringdashboard.core.exceptions.scoped_exception_logger import (
-    ScopedExceptionLogger,
+from flask_monitoringdashboard.core.exceptions.scoped_exception_collector import (
+    ScopedExceptionCollector,
 )
 
 
@@ -133,5 +133,5 @@ def add_database_pruning_schedule(weeks_to_keep, delete_custom_graph_data, **sch
 
 def capture(e: Exception):
     if "scoped_logger" not in g:
-        g.scoped_logger = ScopedExceptionLogger()
+        g.scoped_logger = ScopedExceptionCollector()
     g.scoped_logger.add_user_captured_exc(e)
